requirements.txt
torch
pyyaml==5.1
Pillow
albumentations
opencv-python
torchvision
pyTelegramBotAPI